requirements.txt
pandas>=2.0
numpy<2.0
scipy>=1.10
scikit-surprise>=1.1.3
Flask>=2.3
openai>=1.0.0
python-dotenv>=1.0.0
reportlab==4.0.4